Refactor auth-utils
